demodularized
