+++ docs/source/usage.rst
@@ -9,7 +9,7 @@ The two classes and one function described here have a common signature:
   :samp:`{funcOrClass}(candidate: Any, **kwargs: bool) -> {rtype}`
 
 In normal usage, ``candidate`` must be a :py:class:`str`,
-but we want :func:`hostname.name.is_hostname` to be defined for all inputs,
+but we want :func:`hostname.is_hostname` to be defined for all inputs,
 and we want the others to :exc:`TypeError` only when there is a type problem with the other arguments, while raising some subclass of :class:`hostname.exceptions.HostnameException` on a malformed candidate.
 
 Details on the keyword arguments are described in :ref:`sec-flags` below.
@@ -19,15 +19,15 @@ Details on the keyword arguments are described in :ref:`sec-flags` below.
 The function
 -------------
 
-.. autofunction:: hostname.name.is_hostname
+.. autofunction:: hostname.is_hostname
 
-The function, :func:`hostname.name.is_hostname` returns |True| if and only
+The function, :func:`hostname.is_hostname` returns |True| if and only
 if its argument is a syntactically valid hostname.
 It's behavior can be adjusted slightly with some flags which can be set as keyword arguements.
 
 Additionally it acts as a :py:class:`typing.TypeGuard`,
 allowing type checkers to know that any object which
-passes the test is a :class:`hostname.name.Hostname`.
+passes the test is a :class:`hostname.Hostname`.
 
 
 Classes
@@ -36,9 +36,9 @@ Classes
 Name
 ^^^^^
 
-The :class:`hostname.name.Name` is the substantive class.
+The :class:`hostname.Name` is the substantive class.
 
-.. autoclass:: hostname.name.Name
+.. autoclass:: hostname.Name
     :members:
 
 If ``candidate`` is not a valid hostname, initializing the class will one
@@ -47,9 +47,9 @@ of the :exc:`hostname.exceptions.HostnameException` :doc:`exceptions`.
 Hostname
 ^^^^^^^^
 
-The :class:`hostname.name.Hostname` should be thought of as a subtype of :py:class:`str`.
+The :class:`hostname.Hostname` should be thought of as a subtype of :py:class:`str`.
 
-.. autoclass:: hostname.name.Hostname
+.. autoclass:: hostname.Hostname
 
 If ``candidate`` is not a valid hostname, initializing the class will one
 of the :exc:`hostname.exceptions.HostnameException` :doc:`exceptions`.
